Upgraded to osmdroid 6.0.1

+++ app/src/main/java/org/dyndns/gill_roxrud/frodeg/animalexchange/logic/AnimalGiftManager.java
@@ -7,6 +7,7 @@ import org.dyndns.gill_roxrud.frodeg.animalexchange.InvalidPositionException;
 import org.dyndns.gill_roxrud.frodeg.animalexchange.Point;
 import org.osmdroid.util.GeoPoint;
 
+import java.util.HashSet;
 import java.util.Set;
 
 
@@ -14,6 +15,8 @@ public class AnimalGiftManager {
     private static final int HOR_GIFT_COUNT = 32768;
     private static final int VER_GIFT_COUNT = 16384;
     public static final float GIFT_SIZE_RADIUS = 75.0f; //meters
+    static final double HALF_HOR_GIFT_DEGREE = (AnimalExchangeApplication.HOR_DEGREES/HOR_GIFT_COUNT)/2; //Used for rounding
+    static final double HALF_VER_GIFT_DEGREE = (AnimalExchangeApplication.VER_DEGREES/VER_GIFT_COUNT)/2; //Used for rounding
 
     private int offsetDay = 0;
     private Point<Double> dailyOffset;
@@ -27,23 +30,51 @@ public class AnimalGiftManager {
     public AnimalGiftManager() {
     }
 
-    public Animal requestAnimalGiftT(final Point<Double> pos, final int day) throws InvalidPositionException {
-        return null; //TODO
+    public Animal requestAnimalGift(final Point<Double> pos, final int day) throws InvalidPositionException {
+        double horizontal_pos_rounded = pos.getX() + HALF_HOR_GIFT_DEGREE;
+        if (AnimalExchangeApplication.EAST<=horizontal_pos_rounded) {
+            horizontal_pos_rounded -= AnimalExchangeApplication.HOR_DEGREES;
+        }
+
+        double vertical_pos_rounded = pos.getY() + HALF_VER_GIFT_DEGREE;
+
+        Point<Integer> p = new Point<>(ToHorizontalGift(horizontal_pos_rounded),
+                                       ToVerticalGift(vertical_pos_rounded));
+
+        Point<Double> offset = getOffset(day); //-0.5 to 0.5
+        Point<Double> nearest_gift_pos = new Point<>(FromHorizontalGift(p.getX() + offset.getX()),
+                                                     FromVerticalGift(p.getY() + offset.getY()));
+
+        if (GIFT_SIZE_RADIUS >= CalculateDistance(pos, nearest_gift_pos)) {
+            if (giftsAwarded==null || awardDay!=day) {
+                giftsAwarded = new HashSet<>();
+                awardDay = day;
+            }
+            giftsAwarded.add(ToAnimalGiftKey(p.getX(), p.getY()));
+
+            long distributionValue = AnimalManager.calculateAnimalDistributionValue(p.getX(), p.getY(), day);
+            return GameState.getInstance().getAnimalManager().getAnimalFromDistributionValue(distributionValue);
+        }
+
+        return null;
     }
 
     public boolean isAwardedT(final int key, final int day) {
         try {
+            if (giftsAwarded!=null && awardDay==day) {
+                return giftsAwarded.contains(key);
+            }
+
             AnimalExchangeDBHelper db = GameState.getInstance().getDB();
             if (lastPurgeDay < day) {
                 db.purgeOldAnimalGiftsT(day);
                 lastPurgeDay = day;
             }
 
-            if (giftsAwarded==null || awardDay!=day)
-            {
+            if (giftsAwarded==null || awardDay!=day) {
                 giftsAwarded = db.fetchAwardedGifts(day);
             }
-            return (giftsAwarded==null || giftsAwarded.contains(day));
+            return (giftsAwarded==null || giftsAwarded.contains(key));
         }
         catch(Exception e){
             return true;
@@ -72,15 +103,15 @@ public class AnimalGiftManager {
         return Math.min(value, VER_GIFT_COUNT);
     }
 
-    public static double FromHorizontalBonusGridF(final double x_grid) {
+    public static double FromHorizontalGift(final double x_grid) {
         return AnimalExchangeApplication.WEST + (x_grid/(double) HOR_GIFT_COUNT) * AnimalExchangeApplication.HOR_DEGREES;
     }
 
-    public static double FromVerticalBonusGridF(final double y_grid) {
+    public static double FromVerticalGift(final double y_grid) {
         return AnimalExchangeApplication.MAX_SOUTH_POS + (y_grid/(double) VER_GIFT_COUNT) * AnimalExchangeApplication.VER_ANIMAL_DEGREES;
     }
 
-    public int ToBonusKey(final int x, final int y) throws InvalidPositionException {
+    public int ToAnimalGiftKey(final int x, final int y) throws InvalidPositionException {
         if (VER_GIFT_COUNT <=y || HOR_GIFT_COUNT <=x)
             throw new InvalidPositionException();
 
@@ -104,8 +135,7 @@ public class AnimalGiftManager {
 
     public GeoPoint GeoPointFromGrid(final int x, final int y, final int day) {
         Point<Double> offset = getOffset(day);
-        return new GeoPoint(FromVerticalBonusGridF(y+offset.getY()),
-                            FromHorizontalBonusGridF(x+offset.getX()));
+        return new GeoPoint(FromVerticalGift(y+offset.getY()), FromHorizontalGift(x+offset.getX()));
     }
 
     Point<Double> getOffset(final int day) {
